Added !command usage and edit logging.

+++ bot.py
@@ -50,33 +50,48 @@ async def on_message(message):
 
     if message.content.startswith('!ping'):
         await client.send_message(message.channel, 'Pong')
+
     elif message.content.startswith('!angrave'):
         dir = os.path.dirname(__file__)
         path = 'images/angrave.jpg'
         file_path = os.path.join(dir, path)
         await client.send_file(message.channel, file_path)
+
     elif message.content.startswith('!choice'):
+        if args[1] == 'usage':
+            await client.send_message(message.channel, '```Usage: !choice [first choice, second choice, third choice...]```')
         choices = message.content.split(', ')
         choice = randint(0, len(choices)-1)
-        if choice == 0:
+        if choice == 0 and args[1] != 'usage':
              await client.send_message(message.channel, choices[0][7:])
-        else:
+        elif args[1] != 'usage':
             await client.send_message(message.channel, choices[choice])
+
     elif message.content.startswith('!pepe'):
         await client.send_message(message.channel, ':frog::frog::frog::frog::frog::frog::frog:\n:frog::frog::frog::frog::frog::frog::frog::frog::frog:\n:frog::frog::frog::frog::frog::frog::frog::frog::frog::frog::frog:\n:frog::frog::frog::frog::frog::frog::frog::frog::frog::frog::frog::frog::frog:\n:frog::frog::white_circle:️:black_circle:️:black_circle:️:white_circle:️:frog::frog::frog::white_circle:️:black_circle:️:black_circle:️:white_circle:️\n:frog::white_circle:️:black_circle:️:black_circle:️:white_circle:️:black_circle:️:white_circle:️:frog::white_circle:️:black_circle:️:black_circle:️:white_circle:️:black_circle:️:white_circle:️\n:frog::white_circle:️:black_circle:️:white_circle:️:black_circle:️:black_circle:️:white_circle:️:frog::white_circle:️:black_circle:️:white_circle:️:black_circle:️:black_circle:️:white_circle:️\n:frog::frog::white_circle:️:black_circle:️:white_circle:️:white_circle:️:frog::frog::frog::white_circle:️:black_circle:️:white_circle:️:white_circle:️\n:frog::frog::frog::frog::frog::frog::frog::frog::frog::frog::frog::frog::frog:\n:red_circle::red_circle::frog::frog::frog::frog::frog::frog::frog::frog::frog::frog::frog:\n:frog::red_circle::red_circle::frog::frog::frog::frog::frog::frog::frog::frog::frog:\n:frog::frog::red_circle::red_circle::red_circle::red_circle::red_circle::red_circle::red_circle::red_circle::red_circle::red_circle::red_circle:\n :frog::frog::frog::red_circle::red_circle::red_circle::red_circle::red_circle::red_circle::red_circle::red_circle::red_circle::red_circle:\n:frog::frog::frog::frog::frog::frog::frog::frog::frog::frog::frog:\n:frog::frog::frog::frog::frog::frog::frog::frog::frog::frog:\n:frog::frog::frog::frog::frog::frog::frog::frog::frog:')
+    
     elif message.content.startswith('!xd'):
         await client.send_message(message.channel, ':joy::joy::joy::joy::joy::joy::joy::joy::joy::joy::joy::joy::joy::joy::joy:\n:joy::cool::cool::cool::cool::cool::cool::cool::cool::cool::cool::cool::cool::cool::joy:\n:joy::cool::100::cool::cool::cool::100::cool::100::100::100::cool::cool::cool::joy:\n:joy::cool::100::100::cool::100::100::cool::100::cool::100::100::cool::cool::joy:\n:joy::cool::cool::100::cool::100::cool::cool::100::cool::cool::100::100::cool::joy:\n:joy::cool::cool::100::100::100::cool::cool::100::cool::cool::cool::100::cool::joy:\n:joy::cool::cool::cool::100::cool::cool::cool::100::cool::cool::cool::100::cool::joy:\n:joy::cool::cool::100::100::100::cool::cool::100::cool::cool::cool::100::cool::joy:\n:joy::cool::cool::100::cool::100::cool::cool::100::cool::cool::100::100::cool::joy:\n:joy::cool::100::100::cool::100::100::cool::100::cool::100::100::cool::cool::joy:\n:joy::cool::100::cool::cool::cool::100::cool::100::100::100::cool::cool::cool::joy:\n :joy::cool::cool::cool::cool::cool::cool::cool::cool::cool::cool::cool::cool::cool::joy:\n:joy::joy::joy::joy::joy::joy::joy::joy::joy::joy::joy::joy::joy::joy::joy:')
+    
     elif message.content.startswith('!clean'):
         num = 0
-        if len(args) > 0:
+        if args[1] == 'usage':
+            await client.send_message(message.channel, '```Usage: !clean [number of messages to delete]```')
+        if len(args) > 0 and args[1] != 'usage':
             num = int(args[1])
-        for i in range(num+1):
-            await client.delete_message(client.messages.pop())
+        if args[1] != 'usage':
+            for i in range(num+1):
+                await client.delete_message(client.messages.pop())
         client.delete_message(message)
-    elif message.content.startswith('!fact'):
-        if len(args) > 1:
-            n = int(args[1])
-            await client.send_message(message.channel, factorial(n))
+   
+   elif message.content.startswith('!factorial'):
+        if args[1] == 'usage':
+            await client.send_message(message.channel, '```Usage: !factorial [number to take the factorial of]```')
+        else:
+            if len(args) > 1:
+              n = int(args[1])
+              await client.send_message(message.channel, 'The factorial of ' + str(n) + ' is ' + str(factorial(n)) +'.')
+    
     elif message.content.startswith('!hangman'):
         game_message = ""
         if len(args) > 1:
@@ -89,6 +104,9 @@ async def on_message(message):
 
 @client.event
 async def on_message_delete(message):
-    await client.send_message((discord.Object(id='386360744503017493')), message.author.name + '\'s message \"'+ message.content+'\" was deleted')
+    await client.send_message((discord.Object(id='386360744503017493')), message.author.name + '\'s message \"' + message.content + '\" was deleted')
+@client.event
+async def on_message_edit(before, after):
+    await client.send_message((discord.Object(id='386360744503017493')), before.author.name + '\'s message \"' + before.content + '\" was edited to: \"' + after.content + '\".' )
 
-client.run('Insert-token-here')
+client.run('InsertToken')
